requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
pytz
